Add a programme to process bounding box images.
This is unfinished, but the core logic is complete. The idea is that
the previously very tedious work of defining the bounding boxes of
elements manually by entering the coordinates into the plain text
.settings files is replaced by drawing the bounding box directly
inside the graphics programme as a separate layer and then exporting
that layer as a separate PNG file. The committed programme looks for
parts that are not fully transparent (alpha > 0) and uses them to
compute the bounding box of the element.

This programme is written with the assumption that tilests are used
and processes entire tileset PNG files instead of single element PNGs,
i.e. one can directly define all bounding boxes of a tileset with it
in one go.