Convert tscproc programme from C to C++.
The task it does would be simple enough for a plain C programme, but
parsing the generating XML requires an XML parser. Since the decision
has been made for TSC that its XML parser is going to be Xerces-C,
which is a C++ library, that would cause problem. Effectively tscproc
would have needed to depend on another XML parser like expat or
libxml2 that is plain C. Depending on another XML parser than the main
TSC just for the sake of using plain C is undesired, thus it was
required to convert tscproc to C++. It can now make use of Xerces-C
for parsing its XML as well (not implemented yet).