Add tsc-level-upgrade script.
This script is intended to parse TSC2 levels and output TSC3 levels.
For now, it can convert sprites using TSC2 ground images into TSC3
Ground instances. It requires a mapping table of image pathes to
tileset IDs; this mapping table resides at the end of the script and
needs to be manually created as the tilesets are made from the TSC2
ground images.

Other objects are not yet converted.